requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
pytest>=7.0.0
python-dotenv>=1.0.0
alpaca-trade-api>=3.0.0
schedule>=1.2.0